Fix flakeyness of "test_kill_user_code" This test looks at the logs and checks they are as they should be. However if shepherd crashes in this test then the usercode continues running and continues writing to the logs. This commit uses context managers to harden against this but does not completely solve the problem

+++ app/main.py
@@ -2,49 +2,53 @@ from fastapi import FastAPI
 from fastapi.staticfiles import StaticFiles
 import uvicorn
 import os
-import sys
-import logging
-
+from pathlib import Path
 
 from app.routers import runner_router, upload_router, files_router
 from app.run import runner
 from app.config import config
 from app.upload import increase_max_file_size
+from app.logging import configure_logging
 
-app = FastAPI(
+
+configure_logging()
+
+shepherd = FastAPI(
     title="Shepherd",
     version="0.0.1",
     docs_url="/api_docs",
     redoc_url="/api_redoc",
+    debug=True,
 )
 
-app.include_router(runner_router)
-app.include_router(upload_router)
-app.include_router(files_router)
+shepherd.include_router(runner_router)
+shepherd.include_router(upload_router)
+shepherd.include_router(files_router)
 
-app.mount("/editor", StaticFiles(directory=config.editor_path, html=True), name="editor")
-app.mount("/docs", StaticFiles(directory=config.docs_path, html=True), name="docs")
+shepherd.mount("/editor", StaticFiles(directory=config.editor_path, html=True), name="editor")
+shepherd.mount("/docs", StaticFiles(directory=config.docs_path, html=True), name="docs")
 
 
-@app.on_event("startup")
+@shepherd.on_event("startup")
 def startup_event():
-    increase_max_file_size()
-    logging.getLogger("uvicorn").removeHandler(logging.getLogger("uvicorn").handlers[0])
-    # uvicorn_logging = logging.getLogger("uvicorn")
-    # uvicorn_logging.removeHandler(uvicorn_logging.handlers[0])
-    # logging.add(sys.stdout, colorize=True, format="{time:YYYY-MM-DD at HH:mm:ss} | {level} | {message}")
+    """Setup which is run before the rest of shepherd.
 
-@app.on_event("shutdown")
+    """
+    increase_max_file_size()
+
+
+@shepherd.on_event("shutdown")
 def shutdown_event():
     """Kill any running usercode"""
     runner.shutdown()
-    os.remove(config.usr_fifo_path)
+    if os.path.exists(config.usr_fifo_path) is True:
+        os.remove(config.usr_fifo_path)
 
 
-@app.get("/")
+@shepherd.get("/")
 def root():
     return "Root of shepherd-2"
 
 
 if __name__ == '__main__':
-    uvicorn.run(app, port=8080, host='0.0.0.0')
+    uvicorn.run(shepherd, port=8080, host='0.0.0.0')
